+++ fili-core/src/main/java/com/yahoo/bard/webservice/druid/model/datasource/DataSource.java
@@ -49,7 +49,7 @@ public abstract class DataSource {
         return Collections.unmodifiableSet(
                 getPhysicalTables()
                         .stream()
-                        .map(PhysicalTable::getName)
+                        .map(PhysicalTable::getDruidName)
                         .collect(Collectors.toSet())
         );
     }